pyproject.toml
[project]
name = "dazi"
version = "0.1.0"
description = "A coding assistant powered by LLM agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.4.0",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
    "pydantic>=2.0.0",
    "tiktoken>=0.8.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.27.0",
]

[project.scripts]
dazi = "dazi.__main__:main"

[tool.uv]
package = true
